Add nullability metadata to BXEmulatorConfiguration. Fix some static analyzer warnings.

+++ Other Sources/ADBToolkit/NSString+ADBStringFormatting.m
@@ -71,7 +71,7 @@
 	NSUInteger length = self.length;
 	
 	//We will stuff all our actual lines into this
-	NSMutableArray *wrappedLines = [NSMutableArray arrayWithCapacity: (NSUInteger)ceilf(length / (float)maxLength)];
+	NSMutableArray *wrappedLines = [NSMutableArray arrayWithCapacity: (NSUInteger)ceil(length / (double)maxLength)];
 	
 	//Walk over every line of the string
 	for (NSString *line in self.lineEnumerator)
@@ -110,7 +110,7 @@
 - (NSArray *) _linesWrappedByWordAtLength: (NSUInteger)maxLength
 {
 	NSUInteger length = self.length;
-	NSMutableArray *lines = [NSMutableArray arrayWithCapacity: (NSUInteger)ceilf(length / (float)maxLength)];
+	NSMutableArray *lines = [NSMutableArray arrayWithCapacity: (NSUInteger)ceil(length / (double)maxLength)];
 	
 	//IMPLEMENTATION NOTE: we've already split on linebreaks upstream in componentsSplitAtLineLength:atWordBoundaries,
 	//so we don't check for them again here. It would probably be quicker to do it all in one go here though.
@@ -155,7 +155,7 @@
 	NSUInteger length = self.length;
 	NSUInteger offset = 0;
 	
-	NSMutableArray *lines = [NSMutableArray arrayWithCapacity: (NSUInteger)ceilf(length / (float)maxLength)];
+	NSMutableArray *lines = [NSMutableArray arrayWithCapacity: (NSUInteger)ceil(length / (double)maxLength)];
 	
 	while (offset < length)
 	{
